Added Exception handler methods.

+++ src/main/java/com/stackroute/trackservice/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.trackservice.service;
 
 import com.stackroute.trackservice.domain.Track;
+import com.stackroute.trackservice.exceptions.TrackAlreadyExistsException;
+import com.stackroute.trackservice.exceptions.TrackNotFoundException;
 import com.stackroute.trackservice.repository.TrackRepository;
 import org.springframework.stereotype.Service;
 
@@ -8,43 +10,58 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class TrackServiceImpl implements TrackService
-{
+public class TrackServiceImpl implements TrackService {
     private TrackRepository trackRepository;
 
-    public TrackServiceImpl(TrackRepository trackRepository)
-    {
+    public TrackServiceImpl(TrackRepository trackRepository) {
         this.trackRepository = trackRepository;
     }
 
     @Override
-    public Track saveTrack(Track track)             // Save Track
+    public Track saveTrack(Track track) throws TrackAlreadyExistsException            // Save Track
     {
+        if (trackRepository.existsById(track.getTrackId())) {
+            throw new TrackAlreadyExistsException("Track already exists in DB!");
+        }
         Track savedTrack = trackRepository.save(track);
         return savedTrack;
     }
 
     @Override
-    public Track getTrackbyId(int trackId)          // Get Track by Id
+    public Track getTrackbyId(int trackId) throws TrackNotFoundException          // Get Track by Id
     {
-        Track retrievedTrack = trackRepository.findById(trackId).get();
-        return retrievedTrack;
+        if (trackRepository.existsById(trackId))
+        {
+            Track retrievedTrack = trackRepository.findById(trackId).get();
+            return retrievedTrack;
 
+        } else
+            {
+            throw new TrackNotFoundException("Track not found!");
+        }
     }
 
     @Override
-    public List<Track> getAllTracks()               // Get all Tracks
+    public List<Track> getAllTracks() throws Exception               // Get all Tracks
     {
+        if (trackRepository.findAll().isEmpty())
+            throw new Exception();
         return trackRepository.findAll();
     }
 
     @Override
-    public Optional<Track> deleteTrackById(int trackId)         // Delete Track by Id
+    public Optional<Track> deleteTrackById(int trackId) throws TrackNotFoundException       // Delete Track by Id
     {
-         Optional optional = trackRepository.findById(trackId);
-         if(optional.isPresent())
-             trackRepository.deleteById(trackId);
-         return optional;
+        Optional optional;
+        if (trackRepository.existsById(trackId))
+        {
+            optional = trackRepository.findById(trackId);
+            if (optional.isPresent())
+                trackRepository.deleteById(trackId);
+            return optional;
+        }
+        else
+            throw new TrackNotFoundException("Track not found!");
     }
 
     @Override
@@ -55,8 +72,7 @@ public class TrackServiceImpl implements TrackService
     }
 
     @Override
-    public List<Track> findTrackByName(String trackName)
-    {
+    public List<Track> findTrackByName(String trackName) {
         List<Track> foundTrack = trackRepository.findByTrackName(trackName);
         return (List<Track>) foundTrack;
     }
